pyproject.toml
[tool.poetry]
name = "v4vapp-hive-fastapi-auth"
version = "0.1.0"
description = "A FastAPI based Python API server which includes authentication via a JWT Token signed by a Hive Public Key"
authors = ["Brian of London <brian@v4v.app>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
single-source = "^0.3.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
scrypt = "0.8.20"
beem = {git = "https://github.com/brianoflondon/beem.git"}
mnemonic = "^0.21"
passlib = "^1.7.4"
python-jose = "^3.3.0"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240106"
redis = "^5.0.1"
types-redis = "^4.6.0.20240106"
python-multipart = "^0.0.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
flake8 = "^7.0.0"
black = "^24.1.1"
pytest-asyncio = "^0.23.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
